+++ packages/calculang-js/src/compiler.js
@@ -62,6 +62,8 @@ export default async (entrypoint, options = {}) => {
           .replace(/\\/g, '/')}`; // do I need to add info.loaders?
       },
     },
+    parallelism: 1,
+    //mode: 'none',
     optimization: {
       minimize: false,
       concatenateModules: false, // Do this temporarily because concatenated modules leaks absolute paths - TODO fix!
@@ -105,7 +107,7 @@ export default async (entrypoint, options = {}) => {
         esm: Object.entries(stats.compilation.assets)
           //.entries()
           .filter(([key, value]) => key.indexOf('esm/') != -1)
-          .map(([key, value]) => ({ file: key, source: value.source() })),
+          .map(([key, value]) => ({ file: key, source: key.indexOf('.map') == -1 ? value.source() : value.source() })),
         //a: stats.compilation.assets['a'].source(),
       });
     });