+++ packages/calculang-testcase-models/modularity/join/join_esm/cul_scope_1.mjs
@@ -1,0 +1,5 @@
+import { sales_price } from "./cul_scope_0.mjs";import { rating } from "./cul_scope_0.mjs";import { name } from "./cul_scope_0.mjs";import { item } from "./cul_scope_0.mjs";import { customer } from "./cul_scope_0.mjs";import { order } from "./cul_scope_0.mjs";import { order_, customer_, item_ } from "./cul_scope_2.mjs"; // driving keys
+import { name_, rating_ } from "./cul_scope_3.mjs";
+import { sales_price_ } from "./cul_scope_4.mjs";
+
+export { name_, rating_, sales_price_, order_, customer_, item_ };