+++ packages/calculang-testcase-models/modularity/nums/threeAgain-nomemo.cul.js
@@ -1,0 +1,3 @@
+import { num_ as num$ } from './twoAgain.cul';
+
+export const num = () => 1 + num$();