+++ packages/calculang-testcase-models/modularity/nums/threeAgain_esm/cul_scope_3.cul.js
@@ -1,0 +1,2 @@
+import { num } from './two.cul.js';
+export { num };