+++ packages/calculang-testcase-models/modularity/nums/threeAgain_esm/cul_scope_3.mjs
@@ -1,0 +1,2 @@
+import { num } from "./cul_scope_0.mjs";import { num as num_ } from "./cul_scope_4.mjs";
+export { num_ };