+++ packages/calculang-testcase-models/modularity/nums/threeAgain_esm/cul_scope_5.mjs
@@ -1,0 +1,3 @@
+import { num } from "./cul_scope_0.mjs";import { num_ as num$ } from "./cul_scope_6.mjs";
+
+export const num_ = ({}) => num$({}) + 1;