+++ packages/calculang-testcase-models/modularity/nums/twoAgain_esm/cul_scope_5.mjs
@@ -1,0 +1,1 @@
+import { num } from "./cul_scope_0.mjs";export const num_ = ({}) => 1;