+++ packages/calculang-testcase-models/modularity/nums/two_esm/cul_scope_1.mjs
@@ -1,0 +1,3 @@
+import { num } from "./cul_scope_0.mjs";import { num_ as num$ } from "./cul_scope_2.mjs";
+
+export const num_ = ({}) => num$({}) + 1;