+++ packages/calculang-testcase-models/modularity/nums/use-twoAgain.mjs
@@ -1,0 +1,3 @@
+import { num } from './twoAgain_esm/cul_scope_0.mjs';
+
+console.log(num({}));